+++ __VP/system/boot.php
@@ -1,28 +1,41 @@
 <?php
 
 /*
- * Main BootLoader/BootStrap
+ * Main
+ * BootLoader
+ * BootStrap
  */
 if (!defined('ROOT')) {
-    require_once $_SERVER['ROOT_PATH'] . $_SERVER['ERROR_PATH'];
+    require $_SERVER['ROOT_PATH'] . $_SERVER['ERROR_PATH'];
 }
 
-ob_start();
+if (preg_match_all('/index\.php/', $_SERVER['PHP_SELF']) &&
+        preg_match_all('/index\.php/', $_SERVER['SCRIPT_FILENAME'])) {
 
-/*
- * Getting Configuration
- */
-require_once ROOT . '__VP/system/configure.php';
-require_once ROOT . '__VP/system/get_conf.php';
-require_once ROOT . '__VP/system/conf.php';
 
-/*
- * Getting Controllers
- */
-require_once ROOT . '__VP/controllers/urls.php';
-require_once ROOT . '__VP/controllers/render.php';
-require_once ROOT . '__VP/controllers/hooks.php';
-require_once ROOT . '__VP/controllers/apps.php';
+    define('MAIN', true);
 
-new VP\Controller\apps();
+    ob_start();
+
+    /*
+     * Getting Configuration
+     */
+
+    require ROOT . '__VP/system/configure.php';
+    require ROOT . '__VP/system/get_conf.php';
+    require ROOT . '__VP/system/conf.php';
+
+    /*
+     * Getting Controllers
+     */
+
+    require ROOT . '__VP/controllers/urls.php';
+    require ROOT . '__VP/controllers/render.php';
+    require ROOT . '__VP/controllers/hooks.php';
+    require ROOT . '__VP/controllers/apps.php';
+
+    new VP\Controller\apps();
+} else {
+    require $_SERVER['ROOT_PATH'] . $_SERVER['ERROR_PATH'];
+}
 ?>