+++ _apps/default/view/404.php
@@ -1,9 +1,0 @@
-<?php
-if (!defined('ROOT')) {
-    require_once $_SERVER['ROOT_PATH'] . $_SERVER['ERROR_PATH'];
-}
-?>
-<div class="main">
-    <h1>404 Error</h1>
-    <h1>Page not found</h1>
-</div>